moving pow crate from core/pow to it's own crate
