+++ chain/Cargo.toml
@@ -13,6 +13,7 @@ serde_derive = "~1.0.8"
 time = "^0.1"
 
 grin_core = { path = "../core" }
+grin_pow = { path = "../pow" }
 grin_store = { path = "../store" }
 secp256k1zkp = { path = "../secp256k1zkp" }
