+++ chain/src/chain.rs
@@ -23,8 +23,9 @@ use secp::pedersen::Commitment;
 use core::core::{Block, BlockHeader, Output};
 use core::core::target::Difficulty;
 use core::core::hash::Hash;
-use core::{consensus, genesis, pow};
-use core::pow::MiningWorker;
+use core::{consensus, genesis};
+use pow;
+use pow::MiningWorker;
 use grin_store;
 use pipe;
 use store;
