+++ grin/Cargo.toml
@@ -13,6 +13,7 @@ grin_p2p = { path = "../p2p" }
 grin_pool = { path = "../pool" }
 grin_util = { path = "../util" }
 grin_wallet = { path = "../wallet" }
+grin_pow = { path = "../pow" }
 secp256k1zkp = { path = "../secp256k1zkp" }
 
 cuckoo_miner = { git = "https://github.com/mimblewimble/cuckoo-miner", tag="grin_integration_5"}
