+++ pow/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "grin_pow"
+version = "0.1.0"
+authors = ["Ignotus Peverell <igno.peverell@protonmail.com>"]
+workspace = ".."
+
+[dependencies]
+bitflags = "~0.7.0"
+blake2-rfc = "~0.2.17"
+byteorder = "^0.5"
+num-bigint = "^0.1.35"
+rust-crypto = "^0.2"
+rand = "^0.3"
+serde = "~1.0.8"
+serde_derive = "~1.0.8"
+time = "^0.1"
+lazy_static = "~0.2.8"
+
+grin_core = { path = "../core" }
+
+secp256k1zkp = { path = "../secp256k1zkp" }
