+++ pow/src/siphash.rs
